backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "3.11.11"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "langchain-openai (>=0.3.6,<0.4.0)",
    "qdrant-client (>=1.13.2,<2.0.0)",
    "langchain (>=0.3.19,<0.4.0)",
    "python-decouple (>=3.8,<4.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "websockets (>=15.0,<16.0)",
    "modal (>=0.73.66,<0.74.0)",
    "langchain-qdrant (>=0.2.0,<0.3.0)",
    "langchain-community (>=0.3.18,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
